+++ ui/main/game/galactic_war/gw_play/gw_referee.js
@@ -151,7 +151,7 @@ define(["shared/gw_common"], function (GW) {
       ai.personality.adv_eco_mod_alone * ai.econ_rate;
     // Support for shared armies
     if (ai.character === "Boss") {
-      for (var i = 0; i < ai.landing_policy.length; i++) {
+      for (var i = 0; i < ai.bossCommanders; i++) {
         slotsArray.push({
           ai: true,
           name: ai.name,