+++ ui/main/game/galactic_war/shared/js/gw_faction_3.js
@@ -13,11 +13,6 @@ define([], function () {
           name: "First Seeker Osiris",
           character: "Boss",
           econ_rate: 1,
-          landing_policy: [
-            "off_player_planet",
-            "on_player_planet",
-            "no_restriction",
-          ],
           personality: {
             percent_open_vehicle: 0.5,
             percent_open_bot: 0.5,
@@ -126,11 +121,6 @@ define([], function () {
           name: "First Seeker Osiris",
           character: "Boss",
           econ_rate: 1,
-          landing_policy: [
-            "off_player_planet",
-            "on_player_planet",
-            "no_restriction",
-          ],
           personality: {
             percent_open_vehicle: 0.5,
             percent_open_bot: 0.5,
@@ -239,11 +229,6 @@ define([], function () {
           name: "First Seeker Osiris",
           character: "Boss",
           econ_rate: 1,
-          landing_policy: [
-            "off_player_planet",
-            "on_player_planet",
-            "no_restriction",
-          ],
           personality: {
             percent_open_vehicle: 0.5,
             percent_open_bot: 0.5,
@@ -352,11 +337,6 @@ define([], function () {
           name: "First Seeker Osiris",
           character: "Boss",
           econ_rate: 1,
-          landing_policy: [
-            "off_player_planet",
-            "on_player_planet",
-            "no_restriction",
-          ],
           personality: {
             percent_open_vehicle: 0.5,
             percent_open_bot: 0.5,
@@ -465,11 +445,6 @@ define([], function () {
           name: "First Seeker Osiris",
           character: "Boss",
           econ_rate: 1,
-          landing_policy: [
-            "off_player_planet",
-            "on_player_planet",
-            "no_restriction",
-          ],
           personality: {
             percent_open_vehicle: 0.5,
             percent_open_bot: 0.5,