+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added personalities: Defender, Swarm, and Economist
 - All bosses now use the Pumpkin commander to make them clearly identifiable via their icon
 - Allied and enemy commanders will sometimes change which spawn they use on multi-planet systems
+- The number of commanders in a boss army is now controlled by difficulty
 
 ## v2.0.4 - 2020-02-13